# openapi.yaml
openapi: 3.1.0
info:
  title: ChatGPT on Mac
  description: |
    通过 SSH 操作用户本人的 MacBook。命令是真的在那台电脑上执行，文件是真的被读写。
    把它当成你自己的终端用：需要什么信息就直接跑命令查，不要反复向用户确认琐事。
  version: 0.1.0

servers:
  - url: https://mac.example.com
    description: 生产环境

security:
  - bearerAuth: []

paths:
  /api/exec_command:
    post:
      operationId: exec_command
      summary: 在 MacBook 上执行 shell 命令
      description: |
        在用户的 MacBook 上执行任意 shell 命令，等价于本人在终端里敲。可跑 git、npm、pnpm、
        python、node、docker、make、编译、测试、起服务、看日志，以及 ls、cat、grep、find、
        sed、mkdir、mv、rm 等文件操作，支持多行脚本与管道重定向。
        exitCode 非 0 是正常结果不是故障。输出大时先用 grep/head/tail 收窄。
        长期运行的服务用 nohup ... & 放后台。写代码文件请改用 write_file。
        返回 409 表示高风险操作，需向用户说明并带 confirmToken 重试。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExecRequest"
      responses:
        "200":
          description: 命令已执行完毕（exitCode 非 0 也走这里）
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExecResult"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          description: 命令涉及高风险操作，需要用户确认后带 confirmToken 重试
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConfirmationRequired"
        "503":
          $ref: "#/components/responses/SshUnavailable"

  /api/read_file:
    post:
      operationId: read_file
      summary: 读取 MacBook 上的文件
      description: |
        以二进制安全的方式读取文件（走 SFTP，不经过 shell），不会因编码或特殊字符出乱码。
        读源码、配置、日志用默认 utf8；图片等二进制文件用 base64。
        只需要文件的某几行时，用 exec_command 配 sed -n '100,200p' 更省。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReadFileRequest"
      responses:
        "200":
          description: 读取成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReadFileResult"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/ServerError"
        "503":
          $ref: "#/components/responses/SshUnavailable"

  /api/write_file:
    post:
      operationId: write_file
      summary: 写入 MacBook 上的文件
      description: |
        以二进制安全的方式写文件（走 SFTP，不经过 shell）。内容里的反引号、$、引号、换行
        都不需要转义，父目录不存在会自动创建。写源码和脚本一律用它 —— 用 exec_command
        配 heredoc 会被 shell 解析或提前截断，写进去的和你想写的不是一回事。
        改已有文件：先 read_file 拿到全文，改好再整份写回。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WriteFileRequest"
      responses:
        "200":
          description: 写入成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WriteFileResult"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          description: 目标是敏感文件，需要用户确认后带 confirmToken 重试
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConfirmationRequired"
        "503":
          $ref: "#/components/responses/SshUnavailable"

  /health/ssh:
    get:
      operationId: check_connection
      summary: 检查到 MacBook 的连接是否正常
      description: 在 MacBook 上跑一条最廉价的命令确认链路通。其他工具报 503 时用它定位问题。
      responses:
        "200":
          description: 链路正常
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthResult"
        "503":
          description: 链路不通，通常是反向 SSH 隧道断了
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthResult"

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer

  responses:
    BadRequest:
      description: 参数不合法
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: Bearer token 缺失或错误
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    ServerError:
      description: 服务内部错误
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    SshUnavailable:
      description: 连不上 MacBook，通常是反向 SSH 隧道断开或电脑休眠
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"

  schemas:
    ExecRequest:
      type: object
      required: [command]
      properties:
        command:
          type: string
          description: 要执行的 shell 命令，支持多行脚本、管道、重定向、heredoc
        cwd:
          type: string
          description: 工作目录，支持 ~ 展开，例如 ~/projects/myapp。默认为用户主目录
        timeoutMs:
          type: integer
          description: 超时毫秒数，默认 60000，上限 900000。编译或安装依赖记得调大
        env:
          type: object
          additionalProperties:
            type: string
          description: 追加的环境变量
        stdin:
          type: string
          description: 喂给命令的标准输入
        confirmToken:
          type: string
          description: 上一次返回 409 时给出的确认令牌，原样填回即可放行该命令

    ExecResult:
      type: object
      properties:
        status:
          type: string
          description: 固定为 ok
        stdout:
          type: string
          description: 标准输出，超过上限会在末尾标注截断信息
        stderr:
          type: string
          description: 标准错误
        exitCode:
          type: integer
          nullable: true
          description: 退出码。0 表示成功，非 0 是命令自身的失败，照常分析即可
        signal:
          type: string
          nullable: true
          description: 被信号终止时的信号名
        timedOut:
          type: boolean
          description: 是否因超时被强制终止
        truncated:
          type: boolean
          description: 输出是否被截断
        durationMs:
          type: integer
          description: 实际耗时
        cwd:
          type: string
          description: 实际使用的工作目录

    ReadFileRequest:
      type: object
      required: [path]
      properties:
        path:
          type: string
          description: 文件路径，支持 ~ 展开
        encoding:
          type: string
          enum: [utf8, base64]
          description: 默认 utf8；二进制文件用 base64

    ReadFileResult:
      type: object
      properties:
        status:
          type: string
        path:
          type: string
          description: 展开后的绝对路径
        content:
          type: string
        encoding:
          type: string
        size:
          type: integer
          description: 文件真实字节数
        truncated:
          type: boolean
        note:
          type: string
          description: 被截断时的说明

    WriteFileRequest:
      type: object
      required: [path, content]
      properties:
        path:
          type: string
          description: 目标文件路径，支持 ~ 展开，父目录会自动创建
        content:
          type: string
          description: 文件内容，任意字符都安全，不需要转义
        encoding:
          type: string
          enum: [utf8, base64]
          description: content 的编码，默认 utf8
        append:
          type: boolean
          description: true 追加到末尾，默认 false 覆盖
        confirmToken:
          type: string
          description: 写敏感文件时返回 409 后拿到的确认令牌

    WriteFileResult:
      type: object
      properties:
        status:
          type: string
        path:
          type: string
        bytesWritten:
          type: integer
        created:
          type: boolean
          description: true 表示新建了文件，false 表示覆盖了已有文件

    ConfirmationRequired:
      type: object
      properties:
        status:
          type: string
          description: 固定为 needs_confirmation
        confirmToken:
          type: string
          description: 把这个值原样放进请求的 confirmToken 字段重试即可执行
        expiresInMs:
          type: integer
          description: 令牌有效期
        reasons:
          type: array
          description: 触发确认的具体原因
          items:
            type: object
            properties:
              id:
                type: string
              label:
                type: string
        message:
          type: string
          description: 面向用户的解释，转述给用户后再决定是否重试

    HealthResult:
      type: object
      properties:
        status:
          type: string
        detail:
          type: string
        latencyMs:
          type: integer

    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
          description: 人类可读的原因和修复建议
        requestId:
          type: string